moved some components into a fermentation graph folder for better organization
