+++ frontend/src/app/shared/components/fermentation-graphs/fermentation-yield-graph/fermentation-yield-graph.component.html
